moved csvy functions to model_reader
